NUnite Logos
These images are the logos I created based of the current Northumbria University logo. There are two versions, one in black and one in white. The white one is there it just looks like its not because of the white background. Please feel free to use them if you need them for your subsystems.