NUnite Prototype IIII
This is the fourth development for the NUnite app. 
Changes included: 
Creation for posts is now available by going on the profile and selecting create post.
Can select what type of post you want posted.
Can upload content and filter posts.
Commenting section has been made for each of the posts so now users can view, like and comment on posts. 
Like count has been added to posts in description of post.
Reporting post page has now been created so that users can filter why they are reporting the post and send it off to the admin for investigation.